backend/requirements.txt
# Web Framework
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0

# Database
SQLAlchemy>=2.0,<3.0

# Validation / Settings
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
email-validator>=2.1,<3.0

# Auth / Security (JWT only — password verification is delegated to OS)
python-jose[cryptography]>=3.3,<4.0

# File uploads & CORS
python-multipart>=0.0.9,<1.0

# Environment
python-dotenv>=1.0,<2.0

# HTTP client (OS session verification)
httpx>=0.27,<1.0
